Make resize_board(), set_first_success() and set_per_cell() shared
