+++ minegauler/core/regular/controller.py
@@ -70,68 +70,6 @@ class GameController(_ControllerMixin, GameControllerBase):
         cells = self.game.chord_on_cell(coord)
         self._send_updates(cells)
 
-    def resize_board(self, x_size: int, y_size: int, mines: int) -> None:
-        """See AbstractController."""
-        super().resize_board(x_size=x_size, y_size=y_size, mines=mines)
-        if (
-            x_size == self._opts.x_size
-            and y_size == self._opts.y_size
-            and mines == self._opts.mines
-        ):
-            logger.info(
-                "No resize required as the parameters are unchanged, starting a new game"
-            )
-            self.new_game()
-            return
-
-        logger.info(
-            "Resizing board from %sx%s with %s mines to %sx%s with %s mines",
-            self._opts.x_size,
-            self._opts.y_size,
-            self._opts.mines,
-            x_size,
-            y_size,
-            mines,
-        )
-        self._opts.x_size = x_size
-        self._opts.y_size = y_size
-        self._opts.mines = mines
-
-        self.game = self.game_cls(
-            x_size=self._opts.x_size,
-            y_size=self._opts.y_size,
-            mines=self._opts.mines,
-            per_cell=self._opts.per_cell,
-            lives=self._opts.lives,
-            first_success=self._opts.first_success,
-        )
-        self._send_resize_update()
-
-    def set_first_success(self, value: bool) -> None:
-        """
-        Set whether the first click should be a guaranteed success.
-        """
-        super().set_first_success(value)
-        self._opts.first_success = value
-        if not self.game.state.started():
-            self.game.first_success = value
-
-    def set_per_cell(self, value: int) -> None:
-        """
-        Set the maximum number of mines per cell.
-        """
-        super().set_per_cell(value)
-        if value < 1:
-            raise ValueError(
-                f"Max number of mines per cell must be at least 1, got {value}"
-            )
-        self._opts.per_cell = value
-        # If the game is not started and the minefield is not known then the
-        # new per-cell value should be picked up immediately, and the board
-        # cleared of any flags (e.g. 3-flag cells may no longer be allowed!).
-        if not (self.game.state.started() or self.game.minefield_known):
-            self.new_game()
-
     def load_minefield(self, file: PathLike) -> None:
         """
         Load a minefield from file.
@@ -217,6 +155,7 @@ class CreateController(_ControllerMixin, CreateControllerBase):
     def resize_board(self, x_size: int, y_size: int, mines: int) -> None:
         """Resize the board. The number of mines is ignored."""
         super().resize_board(x_size=x_size, y_size=y_size, mines=mines)
+        # TODO: can this be shared with the game controller?
         if x_size == self._opts.x_size and y_size == self._opts.y_size:
             logger.info(
                 "No resize required as the parameters are unchanged, starting a new game"
